Deduplicate ESP32Reset
