+++ esp-hal/src/lib.rs
@@ -532,36 +532,6 @@ pub mod __macro_implementation {
     pub use xtensa_lx_rt::entry as __entry;
 }
 
-#[cfg(riscv)]
-#[unsafe(export_name = "hal_main")]
-fn hal_main(a0: usize, a1: usize, a2: usize) -> ! {
-    unsafe extern "Rust" {
-        // This symbol will be provided by the user via `#[entry]`
-        fn main(a0: usize, a1: usize, a2: usize) -> !;
-    }
-
-    unsafe extern "C" {
-        static mut __stack_chk_guard: u32;
-    }
-
-    unsafe {
-        let stack_chk_guard = core::ptr::addr_of_mut!(__stack_chk_guard);
-        // we _should_ use a random value but we don't have a good source for random
-        // numbers here
-        stack_chk_guard.write_volatile(esp_config::esp_config_int!(
-            u32,
-            "ESP_HAL_CONFIG_STACK_GUARD_VALUE"
-        ));
-
-        main(a0, a1, a2);
-    }
-}
-
-#[unsafe(export_name = "__stack_chk_fail")]
-unsafe extern "C" fn stack_chk_fail() {
-    panic!("Stack corruption detected");
-}
-
 #[cfg(feature = "unstable")]
 use crate::config::{WatchdogConfig, WatchdogStatus};
 use crate::{
